backend/requirements-prod.txt
# Core Django
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# PostgreSQL
psycopg2-binary==2.9.11

# CORS & Security
django-cors-headers==4.9.0
django-environ==0.12.1

# API Documentation
drf-spectacular==0.31.0

# Production Server
gunicorn==23.0.0
whitenoise==6.9.0

# Database & Caching
redis==7.1.0
channels==4.3.2
channels-redis==4.3.0

# Async Tasks
celery==5.4.0
django-celery-beat==2.6.0

# File Storage
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
Pillow==12.0.0

# HTTP Requests
requests==2.32.5

# Utilities
python-decouple==3.8
python-dotenv==1.2.1
